Fiks for #998
