+++ SharePointFramework/ProjectWebParts/src/components/DynamicMatrix/DynamicMatrix.tsx
@@ -1,16 +1,18 @@
-import React, { FC } from 'react'
-import { DynamicMatrixContext } from './context'
-import styles from './DynamicMatrix.module.scss'
-import { MatrixCell, MatrixCellType, MatrixElement, MatrixHeaderCell } from './MatrixCell'
-import { MatrixRow } from './MatrixRow'
-import { IDynamicMatrixProps } from './types'
+import React, { FC } from 'react';
+import { DynamicMatrixContext } from './context';
+import styles from './DynamicMatrix.module.scss';
+import { MatrixCellType, MatrixElement, MatrixHeaderCell } from './MatrixCell';
+import { MatrixCell } from './MatrixCell/MatrixCell';
+import { MatrixRow } from './MatrixRow';
+import { IDynamicMatrixProps } from './types';
+
 
 export const DynamicMatrix: FC<IDynamicMatrixProps> = (props) => (
   <DynamicMatrixContext.Provider value={{ props }}>
     <div className={styles.dynamicMatrix} style={{ width: props.width }}>
       {props.configuration.map((row, rowIndex) => {
         const cells = row.map((cell, cellIndex) => {
-          const elements = props.getElementsForCell(cell)
+          const elements = props.getElementsForCell(cell);
           switch (cell.cellType) {
             case MatrixCellType.Cell: {
               return (
@@ -19,27 +21,20 @@ export const DynamicMatrix: FC<IDynamicMatrixProps> = (props) => (
                     <MatrixElement key={idx} {...props} />
                   ))}
                 </MatrixCell>
-              )
+              );
             }
             case MatrixCellType.Header: {
               return (
                 <MatrixHeaderCell
                   key={cellIndex}
                   text={cell.cellValue}
-                  className={cell.className}
-                />
-              )
+                  className={cell.className} />
+              );
             }
           }
-        })
-        return <MatrixRow key={rowIndex}>{cells}</MatrixRow>
+        });
+        return <MatrixRow key={rowIndex}>{cells}</MatrixRow>;
       })}
     </div>
   </DynamicMatrixContext.Provider>
-)
-
-export * from './context'
-export * from './MatrixCell'
-export * from './MatrixRow'
-export * from './types'
-export * from './generateMatrixConfiguration'
+);
