+++ SharePointFramework/ProjectWebParts/src/components/DynamicMatrix/MatrixCell/MatrixElement/index.ts
@@ -1,0 +1,1 @@
+export * from './MatrixElement'
