+++ SharePointFramework/ProjectWebParts/src/components/DynamicMatrix/MatrixCell/index.ts
@@ -1,0 +1,4 @@
+export * from './MatrixCell'
+export * from './MatrixHeaderCell'
+export * from './MatrixElement'
+export * from './types'
