+++ SharePointFramework/ProjectWebParts/src/components/DynamicMatrix/index.ts
@@ -1,0 +1,6 @@
+export * from './DynamicMatrix'
+export * from './context'
+export * from './MatrixCell'
+export * from './MatrixRow'
+export * from './types'
+export * from './generateMatrixConfiguration'
