+++ SharePointFramework/ProjectWebParts/src/components/OpportunityMatrix/index.ts
@@ -1,0 +1,2 @@
+export * from './OpportunityMatrix'
+export * from './types'
