+++ SharePointFramework/ProjectWebParts/src/components/OpportunityMatrix/types.ts
@@ -6,7 +6,7 @@ import { IDynamicMatrixProps } from '../DynamicMatrix'
 
 export interface IOpportunityMatrixProps
   extends Omit<HTMLProps<HTMLDivElement>, 'size'>,
-    Pick<IDynamicMatrixProps, 'size' | 'colorScaleConfig' | 'calloutTemplate'> {
+    Pick<IDynamicMatrixProps, 'calloutTemplate'> {
   /**
    * The items to render in the matrix
    */
