+++ SharePointFramework/ProjectWebParts/src/components/ProjectStatus/Sections/UncertaintySection/UncertaintySection.tsx
@@ -7,8 +7,8 @@ import {
 import * as strings from 'ProjectWebPartsStrings'
 import { UserMessage } from 'pp365-shared-library/lib/components/UserMessage'
 import React, { FC } from 'react'
-import { OpportunityMatrix } from '../../../OpportunityMatrix'
-import { RiskMatrix } from '../../../RiskMatrix'
+import { OpportunityMatrix } from 'components/OpportunityMatrix/OpportunityMatrix'
+import { RiskMatrix } from 'components/RiskMatrix/RiskMatrix'
 import { StatusElement } from '../../StatusElement'
 import { useProjectStatusContext } from '../../context'
 import { BaseSection } from '../BaseSection/BaseSection'
