+++ SharePointFramework/ProjectWebParts/src/components/RiskMatrix/index.ts
@@ -1,0 +1,2 @@
+export * from './RiskMatrix'
+export * from './types'
