+++ SharePointFramework/ProjectWebParts/src/webparts/riskMatrix/index.ts
@@ -1,21 +1,20 @@
-import { format } from '@fluentui/react'
 import { get } from '@microsoft/sp-lodash-subset'
 import {
   IPropertyPaneConfiguration,
-  IPropertyPaneField,
   PropertyPaneDropdown,
   PropertyPaneSlider,
   PropertyPaneTextField,
   PropertyPaneToggle
 } from '@microsoft/sp-property-pane'
 import * as strings from 'ProjectWebPartsStrings'
-import { IRiskMatrixProps, RiskMatrix } from 'components/RiskMatrix'
-import ReactDom from 'react-dom'
-import { BaseProjectWebPart } from '../baseProjectWebPart'
-import { UncertaintyElementModel } from '../../models'
-import { IRiskMatrixWebPartData, IRiskMatrixWebPartProps } from './types'
-import SPDataAdapter from '../../data'
+import { IRiskMatrixProps } from 'components/RiskMatrix'
+import { RiskMatrix } from 'components/RiskMatrix/RiskMatrix'
 import _ from 'lodash'
+import ReactDom from 'react-dom'
+import SPDataAdapter from '../../data'
+import { UncertaintyElementModel } from '../../models'
+import { BaseProjectWebPart } from '../baseProjectWebPart'
+import { IRiskMatrixWebPartData, IRiskMatrixWebPartProps } from './types'
 
 export default class RiskMatrixWebPart extends BaseProjectWebPart<IRiskMatrixWebPartProps> {
   private _data: IRiskMatrixWebPartData = {}
@@ -85,55 +84,6 @@ export default class RiskMatrixWebPart extends BaseProjectWebPart<IRiskMatrixWeb
     ReactDom.unmountComponentAtNode(this.domElement)
   }
 
-  /**
-   * Get header label fields for the property pane.
-   */
-  protected get headerLabelFields(): IPropertyPaneField<any>[] {
-    const size = parseInt(this.properties.size ?? '5', 10)
-    const overrideHeaderLabels = PropertyPaneToggle(`overrideHeaderLabels.${size}`, {
-      label: format(strings.OverrideHeadersLabel, size)
-    })
-    if (!get(this.properties, `overrideHeaderLabels.${size}`, false)) {
-      return [overrideHeaderLabels]
-    }
-    const headerLabelFields: IPropertyPaneField<any>[] = []
-    const probabilityHeaders: string[] = [
-      strings.MatrixHeader_VeryHigh,
-      strings.MatrixHeader_High,
-      strings.MatrixHeader_Medium,
-      strings.MatrixHeader_Low,
-      strings.MatrixHeader_VeryLow,
-      strings.MatrixHeader_ExtremelyLow
-    ]
-    const consequenceHeaders: string[] = [
-      strings.MatrixHeader_Insignificant,
-      strings.MatrixHeader_Small,
-      strings.MatrixHeader_Moderate,
-      strings.MatrixHeader_Serious,
-      strings.MatrixHeader_Critical,
-      strings.MatrixHeader_VeryCritical
-    ]
-    for (let i = 0; i < size; i++) {
-      const probabilityHeaderFieldName = `headerLabels.${size}.p${i}`
-      headerLabelFields.push(
-        PropertyPaneTextField(probabilityHeaderFieldName, {
-          label: format(strings.ProbabilityHeaderFieldLabel, i + 1),
-          placeholder: probabilityHeaders[i]
-        })
-      )
-    }
-    for (let i = 0; i < size; i++) {
-      const consequenceHeaderFieldName = `headerLabels.${size}.c${i}`
-      headerLabelFields.push(
-        PropertyPaneTextField(consequenceHeaderFieldName, {
-          label: format(strings.ConsequenceHeaderFieldLabel, i + 1),
-          placeholder: consequenceHeaders[i]
-        })
-      )
-    }
-    return [overrideHeaderLabels, ...headerLabelFields]
-  }
-
   protected getPropertyPaneConfiguration(): IPropertyPaneConfiguration {
     return {
       pages: [
@@ -170,31 +120,30 @@ export default class RiskMatrixWebPart extends BaseProjectWebPart<IRiskMatrixWeb
                   label: strings.MatrixFullWidthLabel
                 }),
                 !this.properties.fullWidth &&
-                  PropertyPaneSlider('width', {
-                    label: strings.WidthFieldLabel,
-                    min: 400,
-                    max: 1000,
-                    value: 400,
-                    showValue: true,
-                    disabled: this.properties.fullWidth
-                  }),
+                PropertyPaneSlider('width', {
+                  label: strings.WidthFieldLabel,
+                  min: 400,
+                  max: 1000,
+                  value: 400,
+                  showValue: true,
+                  disabled: this.properties.fullWidth
+                }),
                 PropertyPaneTextField('calloutTemplate', {
                   label: strings.CalloutTemplateFieldLabel,
                   multiline: true,
                   resizable: true,
                   rows: 8
                 }),
-                !this.properties.useDynamicConfiguration &&
-                  PropertyPaneDropdown('manualConfigurationPath', {
-                    label: strings.ManualConfigurationPathLabel,
-                    options: this._data.configurations.map(({ url: key, title: text }) => ({
-                      key,
-                      text
-                    })),
-                    selectedKey:
-                      this.properties?.manualConfigurationPath ??
-                      this._data.defaultConfiguration?.url
-                  })
+                PropertyPaneDropdown('manualConfigurationPath', {
+                  label: strings.ManualConfigurationPathLabel,
+                  options: this._data.configurations.map(({ url: key, title: text }) => ({
+                    key,
+                    text
+                  })),
+                  selectedKey:
+                    this.properties?.manualConfigurationPath ??
+                    this._data.defaultConfiguration?.url
+                })
               ].filter(Boolean)
             }
           ]
